+++ app/src/main/java/apollo/app/wofang/bll/Sections.java
@@ -1,14 +1,12 @@
 package apollo.app.wofang.bll;
 
-import com.google.gson.Gson;
 import com.google.gson.reflect.TypeToken;
 
 import java.lang.reflect.Type;
 import java.util.ArrayList;
 import java.util.List;
 
-import apollo.app.wofang.data.dao.sqldal.SectionDao;
-import apollo.core.ApolloApplication;
+import apollo.app.wofang.data.dao.sqldal.SectionABC;
 import apollo.data.model.Section;
 import apollo.util.ResUtil;
 
@@ -26,7 +24,7 @@ public class Sections {
 
     private static void remove(List<Section> sections) {
         for(Section s:sections) {
-            SectionDao.getInstance().delete(s.getId());
+            SectionABC.getInstance().delete(s.getId());
         }
     }
 
@@ -38,7 +36,7 @@ public class Sections {
         t = new TypeToken<ArrayList<Section>>(){}.getType();
         sections = ResUtil.readAsset("sections.json", t);
 
-        SectionDao.getInstance().add(sections);
+        SectionABC.getInstance().add(sections);
     }
 
 
@@ -47,9 +45,9 @@ public class Sections {
 
         ss = getRecommendSections();
         for(Section s:ss) {
-            SectionDao.getInstance().delete(s.getId());
+            SectionABC.getInstance().delete(s.getId());
         }
-        SectionDao.getInstance().add(sections);
+        SectionABC.getInstance().add(sections);
     }
 
     public static void updateSubSections(List<Section> sections) {
@@ -57,9 +55,9 @@ public class Sections {
 
         ss = getSubSections();
         for(Section s:ss) {
-            SectionDao.getInstance().delete(s.getId());
+            SectionABC.getInstance().delete(s.getId());
         }
-        SectionDao.getInstance().add(sections);
+        SectionABC.getInstance().add(sections);
     }
 
     public static List<Section> getRecommendSections() {
@@ -73,10 +71,10 @@ public class Sections {
     private static List<Section> getSections(int type) {
         List<Section> ss = null;
 
-        ss = SectionDao.getInstance().getSections(type, 1, Integer.MAX_VALUE).getObjects();
+        ss = SectionABC.getInstance().getSections(type, 1, Integer.MAX_VALUE).getObjects();
         if (ss.size() == 0) {
             init();
-            ss = SectionDao.getInstance().getSections(type, 1, Integer.MAX_VALUE).getObjects();
+            ss = SectionABC.getInstance().getSections(type, 1, Integer.MAX_VALUE).getObjects();
         }
         return ss;
     }
