+++ app/src/main/java/apollo/app/wofang/widget/SearchContentFragment.java
@@ -1,7 +1,39 @@
 package apollo.app.wofang.widget;
 
+import apollo.data.model.Section;
+import apollo.fragments.WebViewBaseFragment;
+
 /**
  * Created by Texel on 2015/9/1.
  */
-public class SearchContentFragment {
+public class SearchContentFragment extends WebViewBaseFragment<Section> {
+    @Override
+    protected String getCacheKey() {
+        return null;
+    }
+
+    @Override
+    protected void sendRequestData() {
+
+    }
+
+    @Override
+    protected void sendRequestData(String url) {
+
+    }
+
+    @Override
+    protected void executeOnLoadDataSuccess(String content) {
+
+    }
+
+    @Override
+    protected void executeOnLoadDataError(String content) {
+
+    }
+
+    @Override
+    protected void onUrlClick(String url) {
+
+    }
 }
